pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vertector-scylladbstore"
version = "1.0.0"
description = "Production-ready ScyllaDB store with vector search capabilities via Qdrant"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Vertector Team", email = "support@vertector.com"}
]
maintainers = [
    {name = "Vertector Team", email = "support@vertector.com"}
]
keywords = [
    "scylladb",
    "qdrant",
    "vector-search",
    "embeddings",
    "database",
    "nosql",
    "semantic-search",
    "langchain",
    "langgraph"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Database",
    "Topic :: Database :: Database Engines/Servers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
    "Typing :: Typed",
]

dependencies = [
    "scylla-driver==3.29.2",
    "qdrant-client>=1.15.1",
    "pydantic>=2.0.0",
    "python-dotenv",
    "tenacity>=9.1.2",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation>=0.41b0",
    "prometheus-client>=0.18.0",
]

[project.urls]
Homepage = "https://github.com/vertector/vertector-scylladbstore"
Documentation = "https://github.com/vertector/vertector-scylladbstore/blob/main/README.md"
Repository = "https://github.com/vertector/vertector-scylladbstore"
Issues = "https://github.com/vertector/vertector-scylladbstore/issues"
Changelog = "https://github.com/vertector/vertector-scylladbstore/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]
langgraph = [
    "langgraph",
    "langchain-google-genai",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (require ScyllaDB/Qdrant)",
    "load: Load tests (slow, test performance)",
    "chaos: Chaos tests (test failure scenarios)",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
